[refactor] 예매 구역 페이지 진입 로직 분리
